toppr_application/requirements.txt
Django (1.10.1)
djangorestframework (3.4.6)
pip (8.1.2)
setuptools (26.1.1)
wheel (0.29.0)
